Create event registration functionality and interagte it with backend, and also add a functionality of user updating the coverImage of their event

+++ eventBackend/src/controllers/event.controllers.js
@@ -8,7 +8,13 @@ import bcrypt from "bcrypt";
 import { Event } from "../models/event.models.js";
 
 const createEvent = asyncHandler(async (req, res) => {
-    const {eventId, eventName, venue, dateTime, description, organiserName, contact1, contact2, organiserEmailId, imageUrl, eventCreater, eventLaunched} = req.body;
+    const {eventName, venue, dateTime, description, organiserName, contact1, contact2, organiserEmailId, eventCreater, eventLaunched} = req.body;
+
+    const username = req.user?.username;
+    if(!username.endsWith("org")) {
+        console.log("only organisers can create events!");
+        throw new ApiError(403, "only organisers can create events!");
+    }
 
     if(!eventName || !venue || !dateTime || isNaN(Date.parse(dateTime)) || !organiserName || !contact1 || !eventCreater || typeof eventLaunched !== "boolean") {
         console.log("Insufficient event details for creation!");
@@ -29,7 +35,6 @@ const createEvent = asyncHandler(async (req, res) => {
             contact1,
             contact2,
             organiserEmailId,
-            imageUrl,
             eventCreator: eventCreater,
             eventLaunched
         });
@@ -51,14 +56,16 @@ const createEvent = asyncHandler(async (req, res) => {
         throw new ApiError(500, "Event creation failed!");
     }
 });
+
 const getUserEvents = asyncHandler(async (req, res) => {
     const username = req.user?.username;
     let userEvents = await Event.find({eventCreator: username}).select("-createdAt -__v -updatedAt").lean();
-    userEvents = userEvents.map(({_id, eventCreator, dateTime, ...rest}) => ({
+    userEvents = userEvents.map(({_id, eventCreator, dateTime, eventImage, ...rest}) => ({
         ...rest,
         eventCreater: eventCreator,
         eventId: _id,
-        dateTime: new Date(dateTime).toISOString().slice(0, 16)
+        dateTime: new Date(dateTime).toISOString().slice(0, 16),
+        imageUrl: eventImage.url
     }));
     // const dateOnly = new Date(userEvents.dateTime).toISOString().split('T')[0];
     // userEvents.dateTime = dateOnly;
@@ -69,7 +76,7 @@ const getUserEvents = asyncHandler(async (req, res) => {
 });
 
 const editEvent = asyncHandler(async (req, res) => {
-    const {eventId, eventName, venue, dateTime, description, organiserName, contact1, contact2, organiserEmailId, imageUrl, eventCreater, eventLaunched} = req.body;
+    const {eventId, eventName, venue, dateTime, description, organiserName, contact1, contact2, organiserEmailId, eventCreater, eventLaunched} = req.body;
     if(!eventName || !venue || !dateTime || isNaN(Date.parse(dateTime)) || !organiserName || !contact1 || !eventCreater || typeof eventLaunched !== "boolean") {
         console.log("Insufficient event details to edit!");
         throw new ApiError(400, "Some fields are needed to be filled!");
@@ -78,6 +85,12 @@ const editEvent = asyncHandler(async (req, res) => {
         console.log("some other user is using some other user's account to edit events!");
         throw new ApiError(401, "Only account owner can edit events on his behalf!");
     }
+    const username = req.user?.username;
+    if(!username.endsWith("org")) {
+        console.log("only organisers can edit events!");
+        throw new ApiError(403, "only organisers can edit events!");
+    }
+
     try {
      let editedEvent = await Event.findByIdAndUpdate(
         eventId,
@@ -91,7 +104,6 @@ const editEvent = asyncHandler(async (req, res) => {
                 contact1,
                 contact2,
                 organiserEmailId,
-                imageUrl,
                 eventCreator: eventCreater,
                 eventLaunched
             }
@@ -118,23 +130,83 @@ const editEvent = asyncHandler(async (req, res) => {
 
 const getlaunchedEvents = asyncHandler(async (req, res) => {
     let launchedEvents = await Event.find({eventLaunched: true}).select("-createdAt -__v -updatedAt").lean();
-    launchedEvents = launchedEvents.map(({_id, eventCreator, dateTime, ...rest}) => ({
+    launchedEvents = launchedEvents.map(({_id, eventCreator, dateTime,eventImage, ...rest}) => ({
         ...rest,
         eventCreater: eventCreator,
         eventId: _id,
-        dateTime: new Date(dateTime).toISOString().slice(0, 16)
+        dateTime: new Date(dateTime).toISOString().slice(0, 16),
+        imageUrl: eventImage.url
     }));
+
     return res
         .status(200)
         .json(new ApiResponse(200, launchedEvents, "successfully retrived launched events!"));
 });
 
+const UpdateEventPhoto = asyncHandler(async(req, res) => {
+    const {eventId} = req.body;
+    const eventDetails = await Event.findById(eventId).select("eventImage eventCreator").lean();
+    const username = req.user?.username;
+    if(username !== eventDetails.eventCreator) {
+        console.log("only organisers can delete events!");
+        throw new ApiError(500, "only organisers can delete events!");
+    }
+    const photoLocalPath = req.file?.path;
+    console.log("photoLocalPath:", photoLocalPath);
+    if(!photoLocalPath) {
+        throw new ApiError(400, "Photo is required");
+    }
+    let photo;
+    try {
+        photo = await uploadOnCloudinary(photoLocalPath);
+        console.log("uploaded photo", photo);
+    } catch (error) {
+        console.log("Error uploading event photo", error);
+        throw new ApiError(400, "failed to upload the event photo");
+    }
+    try {
+        let previousPublicId = eventDetails.eventImage?.public_id;
+        const event = await Event.findByIdAndUpdate(
+            eventId,
+            {
+                $set: {
+                    eventImage: {
+                        url: photo.url,
+                        public_id: photo.public_id
+                    }
+                }
+            },
+            { new: true }
+        ).select("eventImage").lean();
+        if(previousPublicId) {
+            await deleteFromCloudinary(previousPublicId);
+        }
+        return res
+            .status(200)
+            .json(new ApiResponse(200, {imageUrl: event.eventImage.url}, "Event image successfully updated!"));
+    } catch (error) {
+        console.log("eventImage updation failed!", error);
+        if(photo) {
+            await deleteFromCloudinary(photo.public_id);
+        }
+        throw new ApiError(400, "Failed to update the event image!");
+    }
+});
+
 const deleteEvent = asyncHandler(async (req, res) => {
     const {eventId} = req.body;
+
+    const username = req.user?.username;
+    if(!username.endsWith("org")) {
+        console.log("only organisers can delete events!");
+        throw new ApiError(500, "only organisers can delete events!");
+    }
+
     await Event.findByIdAndDelete(eventId);
+
     return res
         .status(200)
         .json(new ApiResponse(200, {}, "Successfully deleted the event!"));
 });
 
-export {createEvent, getUserEvents, editEvent, getlaunchedEvents, deleteEvent };
+export {createEvent, getUserEvents, editEvent, getlaunchedEvents, deleteEvent, UpdateEventPhoto };
